Added network permissions to app. Finally getting an answer from weather API. Hooray

+++ app/src/main/java/ar/edu/itba/pam/travelapp/model/weather/dtos/Temperature.java
@@ -1,17 +1,17 @@
 package ar.edu.itba.pam.travelapp.model.weather.dtos;
 
 public abstract class Temperature {
-    private final Integer value;
+    private final Double value;
     private final String unit;
     private final Integer unitType;
 
-    protected Temperature(Integer value, String unit, Integer unitType) {
+    protected Temperature(Double value, String unit, Integer unitType) {
         this.value = value;
         this.unit = unit;
         this.unitType = unitType;
     }
 
-    public Integer getValue() {
+    public Double getValue() {
         return value;
     }
