+++ app/src/main/java/ar/edu/itba/pam/travelapp/model/weather/dtos/forecast/Minimum.java
@@ -3,7 +3,7 @@ package ar.edu.itba.pam.travelapp.model.weather.dtos.forecast;
 import ar.edu.itba.pam.travelapp.model.weather.dtos.Temperature;
 
 public class Minimum extends Temperature {
-    protected Minimum(Integer value, String unit, Integer unitType) {
+    protected Minimum(Double value, String unit, Integer unitType) {
         super(value, unit, unitType);
     }
 }
