+++ app/src/main/java/ar/edu/itba/pam/travelapp/model/weather/dtos/location/elevation/Metric.java
@@ -3,7 +3,7 @@ package ar.edu.itba.pam.travelapp.model.weather.dtos.location.elevation;
 import ar.edu.itba.pam.travelapp.model.weather.dtos.Temperature;
 
 public class Metric extends Temperature {
-    protected Metric(Integer value, String unit, Integer unitType) {
+    protected Metric(Double value, String unit, Integer unitType) {
         super(value, unit, unitType);
     }
 }
